# backend/src/config/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "E-Commerce MERN API",
    "version": "1.0.0",
    "description": "API documentation for E-Commerce MERN Stack Application (Auto-generated)",
    "contact": {
      "name": "API Support",
      "email": "support@example.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000/api",
      "description": "Development server"
    },
    {
      "url": "https://api.yourdomain.com/api",
      "description": "Production server"
    }
  ],
...
  "paths": {
...
    "/upload/avatar": {
      "post": {
        "tags": [
          "Upload"
        ],
        "summary": "Upload user avatar",
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/upload/review/images": {
      "post": {
        "tags": [
          "Upload"
        ],
        "summary": "Upload review images",
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/upload/review/video": {
      "post": {
        "tags": [
          "Upload"
        ],
        "summary": "Upload review video",
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/upload/delete": {
      "delete": {
        "tags": [
          "Upload"
        ],
        "summary": "Delete uploaded file",
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "publicId": {
                    "example": "any"
                  },
                  "resourceType": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/upload/delete-multiple": {
      "delete": {
        "tags": [
          "Upload"
        ],
        "summary": "Delete multiple uploaded files",
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "publicIds": {
                    "example": "any"
                  },
                  "resourceType": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/upload/signature": {
      "get": {
        "tags": [
          "Upload"
        ],
        "summary": "Get Cloudinary signature for direct upload",
        "description": "",
        "parameters": [
          {
            "name": "folder",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your JWT token"
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}